backend/requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.0
PyJWT==2.8.0
werkzeug==2.3.0
python-dotenv==1.0.0
PyMySQL==1.1.0
mysql-connector-python==8.1.0